move Pixel interface to the api module
